requirements.txt
streamlit

pandas>=1.3.0
numpy>=1.19.3
matplotlib==3.4.3
google-cloud-bigquery==2.28.0
streamlit-echarts>=0.4.0
plotly
streamlit-authenticator